move exceptions to service package
