+++ lib/src/main/java/live/match/api/LiveScoreboardApiImpl.java
@@ -1,8 +1,6 @@
 package live.match.api;
 
-import live.match.service.Match;
-import live.match.service.MatchService;
-import live.match.service.Scoreboard;
+import live.match.service.*;
 
 public class LiveScoreboardApiImpl implements LiveScoreboardApi {
     private final MatchService matchService;
