+++ lib/src/main/java/live/match/service/MatchNotFoundException.java
@@ -1,4 +1,4 @@
-package live.match.api;
+package live.match.service;
 
 public class MatchNotFoundException extends Exception {
     public MatchNotFoundException(String message) {
