+++ lib/src/main/java/live/match/service/MatchServiceImpl.java
@@ -1,9 +1,5 @@
 package live.match.service;
 
-import live.match.api.InvalidMatchStateException;
-import live.match.api.MatchNotFoundException;
-import live.match.api.StartNewMatchException;
-
 import java.util.*;
 import java.util.stream.Stream;
