+++ lib/src/main/java/live/match/service/StartNewMatchException.java
@@ -1,4 +1,4 @@
-package live.match.api;
+package live.match.service;
 
 public class StartNewMatchException extends Exception {
     public StartNewMatchException(String message) {
